LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/container-inner-at-rules-expected.txt


PASS @keyframes is defined regardless of evaluation
PASS @property is defined regardless of evaluation
PASS @layer order respected regardless of evaluation
PASS @font-face is defined regardless of evaluation
FAIL @media works inside @container assert_equals: expected "true" but got ""
FAIL @supports works inside @container assert_equals: expected "true" but got ""
